refactor: extract local path renaming primitives and orchestrator into domain module
Moves domain types (RenameStatus, RenamePathResult, etc.), template-slot
replacement, path-rename helpers, and the file/folder coordinator from
kleinanzeigen_bot/__init__.py into kleinanzeigen_bot/local_path_renaming.py.

The new module has explicit __all__, public names for operations, and no
dependency on KleinanzeigenBot. Bot methods are reduced to thin wrapper
adapters that derive config values.

- Extract replace_template_id_slot, rename_path_if_target_is_free
- Extract rename_local_ad_file_after_id_change (and folder variant)
- Extract rename_referenced_local_image_file_after_id_change (and batch)
- Extract rename_local_ad_file_and_folder_after_id_change coordinator
- Preserve backward-compat underscored re-exports from kleinanzeigen_bot
- Move focused tests to tests/unit/test_local_path_renaming.py
- Update translations.de.yaml for moved function locations
