+++ src/kleinanzeigen_bot/resources/translations.de.yaml
@@ -191,20 +191,6 @@ kleinanzeigen_bot/__init__.py:
     "DONE: (Re-)published %s": "FERTIG: %s (erneut) veröffentlicht"
     "ad": "Anzeige"
 
-  _rename_path_if_target_is_free:
-    "Skipping local %s rename because target already exists: %s": "Lokale %s-Umbenennung übersprungen, da das Ziel bereits existiert: %s"
-    "Could not rename local %s from %s to %s: %s": "Lokale %s konnte nicht von %s nach %s umbenannt werden: %s"
-    "Renamed local %s from %s to %s": "Lokale %s von %s nach %s umbenannt"
-
-  _rename_local_ad_file_after_id_change:
-    "ad file": "Anzeigendatei"
-
-  _rename_referenced_local_image_file_after_id_change:
-    "image file": "Bilddatei"
-
-  _rename_local_ad_folder_after_id_change:
-    "ad folder": "Anzeigenordner"
-
   publish_ad:
     "Publishing ad '%s'...": "Veröffentliche Anzeige '%s'..."
     "Updating ad '%s'...": "Aktualisiere Anzeige '%s'..."
@@ -370,6 +356,21 @@ kleinanzeigen_bot/__init__.py:
     "DONE: Updated [content_hash] in %s": "FERTIG: [content_hash] in %s aktualisiert."
     "Processing %s/%s: '%s' from [%s]...": "Verarbeite %s/%s: '%s' von [%s]..."
     "ad": "Anzeige"
+
+#################################################
+kleinanzeigen_bot/local_path_renaming.py:
+#################################################
+  rename_path_if_target_is_free:
+    "Could not rename local %s from %s to %s: %s": "Lokale %s konnte nicht von %s nach %s umbenannt werden: %s"
+
+  rename_local_ad_file_after_id_change:
+    "ad file": "Anzeigendatei"
+
+  rename_referenced_local_image_file_after_id_change:
+    "image file": "Bilddatei"
+
+  rename_local_ad_folder_after_id_change:
+    "ad folder": "Anzeigenordner"
 
 #################################################
 kleinanzeigen_bot/price_reduction.py:
